Llms: moved (client) vendors inside ../vendors
