+++ src/modules/llms/vendors/azure/azure.vendor.ts
@@ -2,9 +2,9 @@ import { apiAsync } from '~/modules/trpc/trpc.client';
 
 import { AzureIcon } from '~/common/components/icons/AzureIcon';
 
-import { DLLM } from '../store-llms';
-import { IModelVendor } from '../vendors/IModelVendor';
-import { VChatFunctionIn, VChatMessageIn, VChatMessageOrFunctionCallOut, VChatMessageOut } from '../transports/chatGenerate';
+import { DLLM } from '../../store-llms';
+import { IModelVendor } from '../IModelVendor';
+import { VChatFunctionIn, VChatMessageIn, VChatMessageOrFunctionCallOut, VChatMessageOut } from '../../transports/chatGenerate';
 
 import { LLMOptionsOpenAI } from '../openai/openai.vendor';
 import { OpenAILLMOptions } from '../openai/OpenAILLMOptions';
