+++ src/modules/llms/vendors/oobabooga/oobabooga.vendor.ts
@@ -1,6 +1,6 @@
 import { OobaboogaIcon } from '~/common/components/icons/OobaboogaIcon';
 
-import { IModelVendor } from '../vendors/IModelVendor';
+import { IModelVendor } from '../IModelVendor';
 
 import { LLMOptionsOpenAI, ModelVendorOpenAI } from '../openai/openai.vendor';
 import { OpenAILLMOptions } from '../openai/OpenAILLMOptions';
