+++ src/modules/llms/vendors/openai/OpenAISourceSetup.tsx
@@ -11,8 +11,8 @@ import { InlineError } from '~/common/components/InlineError';
 import { Link } from '~/common/components/Link';
 import { settingsCol1Width, settingsGap } from '~/common/theme';
 
-import type { ModelDescriptionSchema } from '../transports/server/server.common';
-import { DLLM, DModelSource, DModelSourceId, useModelsStore, useSourceSetup } from '../store-llms';
+import type { ModelDescriptionSchema } from '../../transports/server/server.common';
+import { DLLM, DModelSource, DModelSourceId, useModelsStore, useSourceSetup } from '../../store-llms';
 import { hasServerKeyOpenAI, isValidOpenAIApiKey, LLMOptionsOpenAI, ModelVendorOpenAI } from './openai.vendor';
 import { openAIModelToModelDescription } from './openai.data';
